+++ pkg/frontend/querymiddleware/results_cache_test.go
@@ -221,6 +221,28 @@ func TestIsRequestCachable(t *testing.T) {
 			request:  &PrometheusRangeQueryRequest{Query: "sum_over_time(rate(metric[1m])[10m:1m] @ end())", Start: 100000, End: 200000, Step: 5},
 			expected: false,
 		},
+		// offset modifier on vector selectors.
+		{
+			name:     "positive offset on vector selector",
+			request:  &PrometheusRangeQueryRequest{Query: "metric offset 1ms", End: 200000, Step: 5},
+			expected: true,
+		},
+		{
+			name:     "negative offset on vector selector",
+			request:  &PrometheusRangeQueryRequest{Query: "metric offset -1ms", End: 125000, Step: 5},
+			expected: false,
+		},
+		// offset modifier on subqueries.
+		{
+			name:     "positive offset on subqueries",
+			request:  &PrometheusRangeQueryRequest{Query: "sum_over_time(rate(metric[1m])[10m:1m] offset 1ms)", Start: 100000, End: 200000, Step: 5},
+			expected: true,
+		},
+		{
+			name:     "negative offset on subqueries",
+			request:  &PrometheusRangeQueryRequest{Query: "sum_over_time(rate(metric[1m])[10m:1m] offset -1ms)", Start: 100000, End: 200000, Step: 5},
+			expected: false,
+		},
 		// On step aligned and non-aligned requests
 		{
 			name:     "request that is step aligned",